Petites rectifs

+++ js/cart.js
@@ -12,7 +12,13 @@ function NewCartRow(product) {
                         <td><input class="quantityInput" type="number" name="numberOfAddedProducts" 
                         value="${product.quantity}" min="1"></td>
                         <td class="totalPricePerProduct">${totalPricePerProduct} €</td>
-                        <td><button type="button" class="remove_product_btn btn btn-danger btn-sm"><svg width="15" aria-hidden="true" focusable="false" data-prefix="far" data-icon="trash-alt" class="svg-inline--fa fa-trash-alt fa-w-14" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M268 416h24a12 12 0 0 0 12-12V188a12 12 0 0 0-12-12h-24a12 12 0 0 0-12 12v216a12 12 0 0 0 12 12zM432 80h-82.41l-34-56.7A48 48 0 0 0 274.41 0H173.59a48 48 0 0 0-41.16 23.3L98.41 80H16A16 16 0 0 0 0 96v16a16 16 0 0 0 16 16h16v336a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V128h16a16 16 0 0 0 16-16V96a16 16 0 0 0-16-16zM171.84 50.91A6 6 0 0 1 177 48h94a6 6 0 0 1 5.15 2.91L293.61 80H154.39zM368 464H80V128h288zm-212-48h24a12 12 0 0 0 12-12V188a12 12 0 0 0-12-12h-24a12 12 0 0 0-12 12v216a12 12 0 0 0 12 12z"></path></svg></button></td>
+                        <td><button type="button" class="remove_product_btn btn btn-danger btn-sm">
+                                <svg width="15" aria-hidden="true" focusable="false" data-prefix="far" data-icon="trash-alt" class="svg-inline--fa fa-trash-alt fa-w-14" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512">
+                                    <path fill="currentColor" d="M268 416h24a12 12 0 0 0 12-12V188a12 12 0 0 0-12-12h-24a12 12 0 0 0-12 12v216a12 12 0 0 0 12 12zM432 80h-82.41l-34-56.7A48 48 0 0 0 274.41 0H173.59a48 48 0 0 0-41.16 23.3L98.41 80H16A16 16 0 0 0 0 96v16a16 16 0 0 0 16 16h16v336a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V128h16a16 16 0 0 0 16-16V96a16 16 0 0 0-16-16zM171.84 50.91A6 6 0 0 1 177 48h94a6 6 0 0 1 5.15 2.91L293.61 80H154.39zM368 464H80V128h288zm-212-48h24a12 12 0 0 0 12-12V188a12 12 0 0 0-12-12h-24a12 12 0 0 0-12 12v216a12 12 0 0 0 12 12z">
+                                    </path>
+                                </svg>
+                            </button>
+                        </td>
                     <tr>`
     document.getElementById(`cartRowsContainer`).innerHTML += cartRow
 }
@@ -56,12 +62,10 @@ for (let i = 0; i < removeProductBtn.length; i++) {
             //console.log(index)
             cartContent.splice(index, 1)
             localStorage.setItem(`cart-key`, JSON.stringify(cartContent))
-
             window.location.reload()
         }
     })
 }
-
 if (removeProductBtn.length === 0) {
     document.getElementById(`emptyCartAlert`).innerHTML = emptyCartAlert
     TOTALPRICE.innerText = `0 €`
@@ -73,13 +77,14 @@ if (removeProductBtn.length === 0) {
 let clearCartBtn = document.getElementsByClassName(`empty_cart__btn`)
 
 clearCartBtn[0].addEventListener(`click`, () => {
-    alert(`Souhaitez-vous vraiment vider tout le contenu de votre panier ?`)
-    let cartRowsContainer = document.getElementById(`cartRowsContainer`)
-    cartRowsContainer.remove()
-    cartBadge.remove()
-    localStorage.clear()
-    document.getElementById(`emptyCartAlert`).innerHTML = emptyCartAlert
-    TOTALPRICE.innerText = `0 €`
+    if (confirm(`Souhaitez-vous vraiment vider tout le contenu de votre panier ?`)) {
+        let cartRowsContainer = document.getElementById(`cartRowsContainer`)
+        cartRowsContainer.remove()
+        cartBadge.remove()
+        localStorage.clear()
+        document.getElementById(`emptyCartAlert`).innerHTML = emptyCartAlert
+        TOTALPRICE.innerText = `0 €`
+    }
 })
 
 // Calcul et affichage du montant total du panier
